+++ lifecycle/livedata-test-junit4/src/main/java/TestCoroutineDispatcherRule.kt
@@ -19,21 +19,21 @@ package it.czerwinski.android.lifecycle.livedata.test.junit4
 
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.ExperimentalCoroutinesApi
-import kotlinx.coroutines.test.TestCoroutineDispatcher
+import kotlinx.coroutines.test.UnconfinedTestDispatcher
 import kotlinx.coroutines.test.resetMain
 import kotlinx.coroutines.test.setMain
 import org.junit.rules.TestWatcher
 import org.junit.runner.Description
 
 /**
- * JUnit4 test rule that swaps main coroutine dispatcher with [TestCoroutineDispatcher].
+ * JUnit4 test rule that swaps main coroutine dispatcher with [UnconfinedTestDispatcher].
  */
 @ExperimentalCoroutinesApi
 class TestCoroutineDispatcherRule : TestWatcher() {
 
     override fun starting(description: Description?) {
         super.starting(description)
-        Dispatchers.setMain(TestCoroutineDispatcher())
+        Dispatchers.setMain(UnconfinedTestDispatcher())
     }
 
     override fun finished(description: Description?) {