+++ lifecycle/livedata-test-junit5/src/main/java/TestCoroutineDispatcherExtension.kt
@@ -19,7 +19,7 @@ package it.czerwinski.android.lifecycle.livedata.test.junit5
 
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.ExperimentalCoroutinesApi
-import kotlinx.coroutines.test.TestCoroutineDispatcher
+import kotlinx.coroutines.test.UnconfinedTestDispatcher
 import kotlinx.coroutines.test.resetMain
 import kotlinx.coroutines.test.setMain
 import org.junit.jupiter.api.extension.AfterEachCallback
@@ -27,13 +27,13 @@ import org.junit.jupiter.api.extension.BeforeEachCallback
 import org.junit.jupiter.api.extension.ExtensionContext
 
 /**
- * JUnit5 extension that swaps main coroutine dispatcher with [TestCoroutineDispatcher].
+ * JUnit5 extension that swaps main coroutine dispatcher with [UnconfinedTestDispatcher].
  */
 @ExperimentalCoroutinesApi
 class TestCoroutineDispatcherExtension : BeforeEachCallback, AfterEachCallback {
 
     override fun beforeEach(context: ExtensionContext?) {
-        Dispatchers.setMain(TestCoroutineDispatcher())
+        Dispatchers.setMain(UnconfinedTestDispatcher())
     }
 
     override fun afterEach(context: ExtensionContext?) {