+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 ### Changed
+- Replace usages of `TestCoroutineDispatcher` with `UnconfinedTestDispatcher`
 - Upgrade Gradle to `7.4.2`
 - Dependencies:
   - Upgrade Kotlin to `1.6.20`